Separated the programs and theory files.
