+++ lib/programs/ternary.dart
