+++ lib/theory/array.dart
