+++ lib/theory/controlBasic.dart
