+++ lib/theory/looping.dart
